refactor: improved directory organization and server shutdown - Restructured the internal directories for better modularity and maintainability. - Enhanced server initialization for more graceful startup. - Implemented improved shutdown procedures to handle in-flight requests gracefully.
