--- backend/cmd/auth-service/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "log"
-
-func main() {
-	log.Fatal("Auth service is not implemented yet.")
-}
