+++ backend/cmd/auth/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "log"
+
+func main() {
+	log.Fatal("Auth service is not implemented yet.")
+}
