+++ backend/internal/common/handlers/encode.go
@@ -1,0 +1,25 @@
+package handlers
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+)
+
+func encodeJSON(w http.ResponseWriter, status int, data interface{}) error {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(status)
+
+	if err := json.NewEncoder(w).Encode(data); err != nil {
+		return fmt.Errorf("failed to encode JSON response: %w", err)
+	}
+
+	return nil
+}
+
+func decodeJSON(r *http.Request, dst interface{}) error {
+	if err := json.NewDecoder(r.Body).Decode(dst); err != nil {
+		return fmt.Errorf("failed to decode JSON request: %w", err)
+	}
+	return nil
+}
