--- backend/internal/common/middleware/auth_middleware.go
@@ -1,13 +1,0 @@
-package middleware
-
-import (
-	"log"
-	"net/http"
-)
-
-func AuthMiddleware(next HandleFunc) HandleFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		log.Printf("Request received: %s\n", r.RequestURI)
-		next(w, r)
-	}
-}
