+++ frontend/src/renderer/src/env.d.ts
@@ -1,6 +1,6 @@
 /// <reference types="vite/client" />
 interface ImportMetaEnv {
-    readonly VITE_TMP_ENV: string;
+    readonly VITE_WEBSOCKET_URL: string;
 }
 
 interface ImportMeta {
