+++ frontend/src/renderer/src/features/Auth/hooks/useAuth.ts
@@ -1,0 +1,13 @@
+import { useContext } from 'react';
+import { AuthContext, AuthContextType } from '../contexts/AuthContext';
+
+const useAuth = (): AuthContextType => {
+    const authContext = useContext(AuthContext);
+
+    if (!authContext) {
+        throw new Error('useAuth must be used with an auth provider');
+    }
+    return authContext;
+};
+
+export default useAuth;
