+++ frontend/src/renderer/src/main.tsx
@@ -3,7 +3,7 @@ import './assets/main.css';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import { AuthProvider } from './features/auth/AuthProvider';
+import { AuthProvider } from './features/Auth/contexts/AuthProvider';
 import { HashRouter } from 'react-router-dom';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
